requirements.txt
bentoml>=1.0.0
xgboost>=1.6.0
scikit-learn>=1.0.0
numpy>=1.20.0
requests>=2.25.0
html2text>=2020.1.16
firecrawl>=0.1.0
pyyaml>=6.0.0
